[#1745] Move marker types into new marker module
