+++ iceoryx2-cli/iox2-service/src/command/replay.rs
@@ -19,7 +19,7 @@ use crate::cli::ReplayOptions;
 use crate::command::get_pubsub_service_types;
 use anyhow::Result;
 use iceoryx2::prelude::*;
-use iceoryx2::service::builder::{CustomHeaderMarker, CustomPayloadMarker};
+use iceoryx2::service::marker::{CustomHeaderMarker, CustomPayloadMarker};
 use iceoryx2::service::static_config::message_type_details::TypeVariant;
 use iceoryx2_cli::Format;
 use iceoryx2_userland_record_and_replay::prelude::*;
