+++ iceoryx2-ffi/python/src/sample_mut_uninit.rs
@@ -12,7 +12,7 @@
 
 use core::mem::MaybeUninit;
 
-use iceoryx2::service::builder::{CustomHeaderMarker, CustomPayloadMarker};
+use iceoryx2::service::marker::{CustomHeaderMarker, CustomPayloadMarker};
 use iceoryx2_log::fatal_panic;
 use pyo3::prelude::*;
