+++ iceoryx2/conformance-tests/src/service_publish_subscribe.rs
@@ -26,7 +26,7 @@ pub mod service_publish_subscribe {
     use iceoryx2::prelude::{AllocationStrategy, *};
     use iceoryx2::service::builder::publish_subscribe::PublishSubscribeCreateError;
     use iceoryx2::service::builder::publish_subscribe::PublishSubscribeOpenError;
-    use iceoryx2::service::builder::{CustomHeaderMarker, CustomPayloadMarker};
+    use iceoryx2::service::marker::{CustomHeaderMarker, CustomPayloadMarker};
     use iceoryx2::service::messaging_pattern::MessagingPattern;
     use iceoryx2::service::port_factory::publisher::PortFactoryPublisher;
     use iceoryx2::service::static_config::message_type_details::{TypeDetail, TypeVariant};
