+++ iceoryx2/src/pending_response.rs
@@ -69,7 +69,7 @@ use crate::port::client::ClientSharedState;
 use crate::port::details::chunk::Chunk;
 use crate::port::details::chunk_details::ChunkDetails;
 use crate::raw_sample::RawSample;
-use crate::service::builder::CustomPayloadMarker;
+use crate::service::marker::CustomPayloadMarker;
 use crate::{port::ReceiveError, request_mut::RequestMut, response::Response, service};
 
 /// Represents an active connection to all [`Server`](crate::port::server::Server)
