+++ iceoryx2/src/port/client.rs
@@ -95,6 +95,7 @@ use iceoryx2_cal::{
 use iceoryx2_log::{fail, fatal_panic, warn};
 
 use crate::active_request::RequestId;
+use crate::service::marker::{CustomHeaderMarker, CustomPayloadMarker};
 use crate::{
     identifiers::UniqueClientId,
     pending_response::PendingResponse,
@@ -108,7 +109,6 @@ use crate::{
     request_mut_uninit::RequestMutUninit,
     service::{
         self,
-        builder::{CustomHeaderMarker, CustomPayloadMarker},
         dynamic_config::request_response::{ClientDetails, ServerDetails},
         header,
         naming_scheme::data_segment_name,
