+++ iceoryx2/src/port/publisher.rs
@@ -133,9 +133,9 @@ use crate::prelude::BackpressureStrategy;
 use crate::raw_sample::RawSampleMut;
 use crate::sample_mut::SampleMut;
 use crate::sample_mut_uninit::SampleMutUninit;
-use crate::service::builder::{CustomHeaderMarker, CustomPayloadMarker};
 use crate::service::dynamic_config::publish_subscribe::{PublisherDetails, SubscriberDetails};
 use crate::service::header::publish_subscribe::Header;
+use crate::service::marker::{CustomHeaderMarker, CustomPayloadMarker};
 use crate::service::naming_scheme::data_segment_name;
 use crate::service::port_factory::publisher::{LocalPublisherConfig, PortFactoryPublisher};
 use crate::service::static_config::message_type_details::TypeVariant;
