+++ iceoryx2/src/port/reader.rs
@@ -39,9 +39,9 @@ use crate::constants::MAX_BLACKBOARD_KEY_SIZE;
 use crate::identifiers::UniqueReaderId;
 use crate::port::port_name::PortName;
 use crate::prelude::EventId;
-use crate::service::builder::CustomKeyMarker;
 use crate::service::builder::blackboard::{BlackboardResources, KeyMemory};
 use crate::service::dynamic_config::blackboard::ReaderDetails;
+use crate::service::marker::CustomKeyMarker;
 use crate::service::port_factory::reader::ReaderConfig;
 use crate::service::static_config::message_type_details::{TypeDetail, TypeVariant};
 use crate::service::{self, SharedServiceState};
