+++ iceoryx2/src/port/writer.rs
@@ -44,9 +44,9 @@ use crate::constants::MAX_BLACKBOARD_KEY_SIZE;
 use crate::identifiers::UniqueWriterId;
 use crate::port::port_name::PortName;
 use crate::prelude::EventId;
-use crate::service::builder::CustomKeyMarker;
 use crate::service::builder::blackboard::{BlackboardResources, KeyMemory};
 use crate::service::dynamic_config::blackboard::WriterDetails;
+use crate::service::marker::CustomKeyMarker;
 use crate::service::port_factory::writer::WriterConfig;
 use crate::service::static_config::message_type_details::{TypeDetail, TypeVariant};
 use crate::service::{self, SharedServiceState};
